Use sqrtf for float values (has an actual speed gain).
git-svn-id: svn+ssh://svn.gna.org/svn/warzone/trunk@3934 4a71c877-e1ca-e34f-864e-861f7616d084

+++ src/scriptfuncs.c
@@ -10953,7 +10953,7 @@ BOOL scrSqrt(void)
 		return FALSE;
 	}
 
-	scrFunctionResult.v.fval = sqrt(fArg);
+	scrFunctionResult.v.fval = sqrtf(fArg);
 	if (!stackPushResult(VAL_FLOAT, &scrFunctionResult))
 	{
 		return FALSE;
